requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
streamlit-drawable-canvas==0.9.3
numpy==1.24.3
python-magic-bin==0.4.14